release 2023.03.001

+++ src/workbench/components/rolemenumanager.ts
@@ -15,6 +15,7 @@ import {modal} from "../../services/modal.service";
 import {RoleMenuManagerEditRoleModal} from "./rolemenumanagereditrolemodal";
 import {switchMap} from "rxjs";
 import {CdkDragDrop, moveItemInArray} from "@angular/cdk/drag-drop";
+import {configurationService} from "../../services/configuration.service";
 
 
 
@@ -52,6 +53,7 @@ export class RoleMenuManager implements OnInit {
                 private modal: modal,
                 private toast: toast,
                 private cdRef: ChangeDetectorRef,
+                private configurationService: configurationService,
                 public injector: Injector) {
     }
 
@@ -74,6 +76,7 @@ export class RoleMenuManager implements OnInit {
 
         this.backend.postRequest(`configuration/configurator/${table}/${roleModule.id}`, null, {config: data}).subscribe({
             next: () => {
+                this.configurationService.reloadTaskData('rolemodules');
                 this.toast.sendToast('LBL_DATA_SAVED', 'success');
             }
         });
@@ -88,7 +91,8 @@ export class RoleMenuManager implements OnInit {
         viewProvider.view.setViewMode();
         const table = roleModule.scope == 'custom' ? 'sysuicustomrolemodules' : 'sysuirolemodules';
         this.backend.deleteRequest(`configuration/configurator/${table}/${roleModule.id}`).subscribe(() => {
-            this.toast.sendToast('MSG_SUCCESSFULLY_DELETED', 'success')
+            this.configurationService.reloadTaskData('rolemodules');
+            this.toast.sendToast('MSG_SUCCESSFULLY_DELETED', 'success');
         });
         this.roleModules = this.roleModules.filter(id => id.id != roleModule.id);
         this.saveSequence();
@@ -109,7 +113,7 @@ export class RoleMenuManager implements OnInit {
         };
 
         this.roleModules.push(roleModule);
-
+        this.filterModules = this.roleModules.map(rm=>rm.module);
 
         // triggering change detection to find changes in order for angular to render edit view
         this.cdRef.detectChanges();
@@ -168,15 +172,22 @@ export class RoleMenuManager implements OnInit {
      * @param id
      */
     public deleteRole(id: string) {
+        this.modal.confirm('MSG_DELETE_RECORD', 'MSG_DELETE_RECORD').subscribe({
+            next: (res) => {
+                if (res) {
+                    const role = this.roles.find(r => r.id == id);
 
-        const role = this.roles.find(r => r.id == id);
+                    const table = role.scope == 'global' ? 'sysuiroles' : 'sysuicustomroles';
 
-        const table = role.scope == 'global' ? 'sysuiroles' : 'sysuicustomroles';
-
-        this.backend.deleteRequest(`configuration/configurator/${table}/${role.id}`).subscribe(() => {
-            this.toast.sendToast('MSG_SUCCESSFULLY_DELETED', 'success')
-        });
-        this.roles = this.roles.filter(id => id.id != role.id);
+                    this.backend.deleteRequest(`configuration/configurator/${table}/${role.id}`).subscribe(() => {
+                        this.configurationService.reloadTaskData('roles');
+                        this.configurationService.reloadTaskData('sysroles');
+                        this.toast.sendToast('MSG_SUCCESSFULLY_DELETED', 'success');
+                    });
+                    this.roles = this.roles.filter(id => id.id != role.id);
+                }
+            }
+        })
     }
 
     /**
@@ -270,7 +281,9 @@ export class RoleMenuManager implements OnInit {
         });
 
         if (globalEntries.length > 0) {
-            this.backend.postRequest(`configuration/configurator/sysuirolemodules`, null, {config: globalEntries});
+            this.backend.postRequest(`configuration/configurator/sysuirolemodules`, null, {config: globalEntries}).subscribe(() => {
+                this.configurationService.reloadTaskData('rolemodules');
+            });
         }
 
         const customEntries = this.roleModules.filter(e => e.scope == 'custom').map(e => {
@@ -280,7 +293,9 @@ export class RoleMenuManager implements OnInit {
         });
 
         if (customEntries.length > 0) {
-            this.backend.postRequest(`configuration/configurator/sysuicustomrolemodules`, null, {config: customEntries});
+            this.backend.postRequest(`configuration/configurator/sysuicustomrolemodules`, null, {config: customEntries}).subscribe(() => {
+                this.configurationService.reloadTaskData('rolemodules');
+            });
         }
     }
 }
